proxyserver/requirements.txt
flask
waitress
fpdf
werkzeug
pillow
pyyaml
